+++ src/global/images.d.ts
